project/api/project/requirements/base.txt
# Base python packages for project to run

# python
celery==3.1.13
elasticsearch==1.2.0
facebook-sdk
geopy==1.1.2
hiredis==0.1.4
pillow==2.5.2
psycopg2==2.5.3
pycountry==1.8
pyelasticsearch==0.7.1
python-memcached
redis==2.10.3
requests==2.3.0
requests-oauthlib==0.4.1
smartypants==1.8.6


# django related
django==1.6.5
django-allauth==0.18.0
django-autoslug==1.7.2
django-configurations==0.8
django-celery==3.1.10
django-extensions==1.3.10
django-endless-pagination==2.0
django-haystack==2.2.0
django-imagekit==3.2.2
django-model-utils==2.2
django-multiselectfield==0.1.2
django-phonenumber-field==0.6
django-post-office==1.0.0
django-taggit==0.12.1
django-widget-tweaks==1.3


# using git
-e git+git://github.com/comoga/django-constance#egg=django-constance
